gh-106320: Remove private _PyCrossInterpreterData API
Move private _PyCrossInterpreterData API to the internal C API: add a
new pycore_cross_interp.h header file.

Mark _PyCrossInterpreterData_Init() and
_PyCrossInterpreterData_Lookup() functions are static.

Move get_crossinterp_data() and restore_crossinterp_data() from
_testcapi to _testinternalcapi.
